update analyze entities
